Refactored acceptance test DSL
Acceptance test DSL is getting a little out of control so moved a few things around...

Reduces the duplication of code, does mean some options might be invalid under some of the paths, but hopefully overall a fair amount cleaner.
