+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/ExternalMetricPublisherAcceptanceTest.java
@@ -16,7 +16,7 @@ package tech.pegasys.teku.test.acceptance;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.test.acceptance.dsl.AcceptanceTestBase;
 import tech.pegasys.teku.test.acceptance.dsl.ExternalMetricNode;
-import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
+import tech.pegasys.teku.test.acceptance.dsl.TekuBeaconNode;
 
 public class ExternalMetricPublisherAcceptanceTest extends AcceptanceTestBase {
   private static final int ACTIVE_VALIDATOR_COUNT = 8;
@@ -27,7 +27,7 @@ public class ExternalMetricPublisherAcceptanceTest extends AcceptanceTestBase {
     ExternalMetricNode externalMetricNode = createExternalMetricNode();
     externalMetricNode.start();
 
-    final TekuNode tekuNode =
+    final TekuBeaconNode tekuNode =
         createTekuNode(
             config ->
                 config
