+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/SyncAcceptanceTest.java
@@ -17,18 +17,18 @@ import java.util.function.Consumer;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.test.acceptance.dsl.AcceptanceTestBase;
-import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
-import tech.pegasys.teku.test.acceptance.dsl.TekuNode.Config;
+import tech.pegasys.teku.test.acceptance.dsl.TekuBeaconNode;
+import tech.pegasys.teku.test.acceptance.dsl.TekuBeaconNode.Config;
 
 public class SyncAcceptanceTest extends AcceptanceTestBase {
 
   @Test
   public void shouldSyncToNodeWithGreaterFinalizedEpoch() throws Exception {
-    final TekuNode primaryNode = createTekuNode(Config::withRealNetwork);
+    final TekuBeaconNode primaryNode = createTekuNode(Config::withRealNetwork);
 
     primaryNode.start();
     UInt64 genesisTime = primaryNode.getGenesisTime();
-    final TekuNode lateJoiningNode =
+    final TekuBeaconNode lateJoiningNode =
         createTekuNode(configureLateJoiningNode(primaryNode, genesisTime.intValue()));
     primaryNode.waitForNewFinalization();
 
@@ -38,7 +38,7 @@ public class SyncAcceptanceTest extends AcceptanceTestBase {
   }
 
   private Consumer<Config> configureLateJoiningNode(
-      final TekuNode primaryNode, final int genesisTime) {
+      final TekuBeaconNode primaryNode, final int genesisTime) {
     return c ->
         c.withGenesisTime(genesisTime)
             .withRealNetwork()
