+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/validatorslashing/MultiPeersSingleProcessGossipAcceptanceTest.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import tech.pegasys.teku.bls.BLSKeyPair;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.SpecMilestone;
-import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
+import tech.pegasys.teku.test.acceptance.dsl.TekuBeaconNode;
 
 /**
  * Running 2 nodes with VC/BN running in single processes. <br>
@@ -41,7 +41,7 @@ public class MultiPeersSingleProcessGossipAcceptanceTest
     final int genesisTime = timeProvider.getTimeInSeconds().plus(10).intValue();
     final UInt64 altairEpoch = UInt64.valueOf(100);
 
-    final TekuNode firstTekuNode =
+    final TekuBeaconNode firstTekuNode =
         createTekuNode(
             config ->
                 configureNode(config, genesisTime, network)
@@ -52,7 +52,7 @@ public class MultiPeersSingleProcessGossipAcceptanceTest
 
     firstTekuNode.waitForEpochAtOrAbove(1);
 
-    final TekuNode secondTekuNode =
+    final TekuBeaconNode secondTekuNode =
         createTekuNode(
             config ->
                 configureNode(config, genesisTime, network)
