+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/validatorslashing/MultiPeersStandAloneVcBlocksAcceptanceTest.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import tech.pegasys.teku.bls.BLSKeyPair;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.SpecMilestone;
-import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
+import tech.pegasys.teku.test.acceptance.dsl.TekuBeaconNode;
 import tech.pegasys.teku.test.acceptance.dsl.TekuValidatorNode;
 
 /**
@@ -41,7 +41,7 @@ public class MultiPeersStandAloneVcBlocksAcceptanceTest
     final int genesisTime = timeProvider.getTimeInSeconds().plus(10).intValue();
     final UInt64 altairEpoch = UInt64.valueOf(100);
 
-    final TekuNode firstTekuNode =
+    final TekuBeaconNode firstTekuNode =
         createTekuNode(
             config ->
                 configureNode(config, genesisTime, network)
@@ -65,7 +65,7 @@ public class MultiPeersStandAloneVcBlocksAcceptanceTest
         slashedValidatorKeyPair.getSecretKey(),
         slashingEventType);
 
-    final TekuNode secondBeaconNode =
+    final TekuBeaconNode secondBeaconNode =
         createTekuNode(
             config ->
                 configureNode(config, genesisTime, network)
@@ -81,7 +81,7 @@ public class MultiPeersStandAloneVcBlocksAcceptanceTest
                     .withValidatorApiEnabled()
                     .withStopVcWhenValidatorSlashedEnabled()
                     .withInteropValidators(32, 32)
-                    .withBeaconNode(secondBeaconNode));
+                    .withBeaconNodes(secondBeaconNode));
 
     secondBeaconNode.start();
