+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/validatorslashing/MultiPeersStandAloneVcGossipAcceptanceTest.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import tech.pegasys.teku.bls.BLSKeyPair;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.SpecMilestone;
-import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
+import tech.pegasys.teku.test.acceptance.dsl.TekuBeaconNode;
 import tech.pegasys.teku.test.acceptance.dsl.TekuValidatorNode;
 
 /**
@@ -42,7 +42,7 @@ public class MultiPeersStandAloneVcGossipAcceptanceTest
     final int genesisTime = timeProvider.getTimeInSeconds().plus(10).intValue();
     final UInt64 altairEpoch = UInt64.valueOf(100);
 
-    final TekuNode firstTekuNode =
+    final TekuBeaconNode firstTekuNode =
         createTekuNode(
             config ->
                 configureNode(config, genesisTime, network)
@@ -53,7 +53,7 @@ public class MultiPeersStandAloneVcGossipAcceptanceTest
 
     firstTekuNode.waitForEpochAtOrAbove(1);
 
-    final TekuNode secondBeaconNode =
+    final TekuBeaconNode secondBeaconNode =
         createTekuNode(
             config ->
                 configureNode(config, genesisTime, network)
@@ -65,11 +65,11 @@ public class MultiPeersStandAloneVcGossipAcceptanceTest
         createValidatorNode(
             config ->
                 config
-                    .withNetwork("auto")
                     .withValidatorApiEnabled()
+                    .withNetwork("auto")
                     .withStopVcWhenValidatorSlashedEnabled()
                     .withInteropValidators(0, 32)
-                    .withBeaconNode(secondBeaconNode));
+                    .withBeaconNodes(secondBeaconNode));
 
     secondBeaconNode.start();
