+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/validatorslashing/SinglePeerAcceptanceTest.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import tech.pegasys.teku.bls.BLSKeyPair;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.SpecMilestone;
-import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
+import tech.pegasys.teku.test.acceptance.dsl.TekuBeaconNode;
 
 /**
  * Running a single node with BN/VC running in a single process. The slashing event is sent to the
@@ -35,7 +35,7 @@ public class SinglePeerAcceptanceTest extends ValidatorSlashingDetectionAcceptan
     final int genesisTime = timeProvider.getTimeInSeconds().plus(10).intValue();
     final UInt64 altairEpoch = UInt64.valueOf(100);
 
-    final TekuNode tekuNode =
+    final TekuBeaconNode tekuNode =
         createTekuNode(
             config ->
                 configureNode(config, genesisTime, network)
