+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/validatorslashing/SinglePeerStandAloneVcAcceptanceTest.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import tech.pegasys.teku.bls.BLSKeyPair;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.SpecMilestone;
-import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
+import tech.pegasys.teku.test.acceptance.dsl.TekuBeaconNode;
 import tech.pegasys.teku.test.acceptance.dsl.TekuValidatorNode;
 
 /**
@@ -35,7 +35,7 @@ public class SinglePeerStandAloneVcAcceptanceTest extends ValidatorSlashingDetec
     final int genesisTime = timeProvider.getTimeInSeconds().plus(10).intValue();
     final UInt64 altairEpoch = UInt64.valueOf(100);
 
-    final TekuNode beaconNode =
+    final TekuBeaconNode beaconNode =
         createTekuNode(
             config -> configureNode(config, genesisTime, network).withAltairEpoch(altairEpoch));
 
@@ -46,7 +46,7 @@ public class SinglePeerStandAloneVcAcceptanceTest extends ValidatorSlashingDetec
                     .withNetwork("auto")
                     .withStopVcWhenValidatorSlashedEnabled()
                     .withInteropValidators(0, 32)
-                    .withBeaconNode(beaconNode));
+                    .withBeaconNodes(beaconNode));
 
     beaconNode.start();
     validatorClient.start();
